Add sticky scrolling behavior to both homepage columns.
