+++ client/homescreen/style.scss
@@ -50,7 +50,7 @@
 
 .woocommerce-homescreen-column {
 	width: 662px;
-	top: 130px;
+	top: 100px;
 	margin: 0 auto;
 	align-self: flex-start;
